+++ src/main/java/com/buaa/blockchain/message/Message.java
@@ -21,8 +21,8 @@ public class Message implements Serializable {
     private String receiverAddress;
     private String topic;
     private String nodeName;
-    private int height;
-    private int round;
+    private Long height;
+    private Long round;
     private Boolean vote;
     private Block block;
     private List<Block> blockList;
@@ -32,7 +32,7 @@ public class Message implements Serializable {
     /**
      * 无topic，一般用于共识协议的发起
      * */
-    public Message(String nodeName,int height,int round,Block block){
+    public Message(String nodeName,Long height,Long round,Block block){
         this.topic = "";
         this.nodeName = nodeName;
         this.height = height;
@@ -44,7 +44,7 @@ public class Message implements Serializable {
     /**
      * 注明topic
      * */
-    public Message(String topic,String nodeName,int height,int round,Boolean vote,Block block){
+    public Message(String topic,String nodeName,Long height,Long round,Boolean vote,Block block){
         this.topic = topic;
         this.nodeName = nodeName;
         this.height = height;
@@ -56,7 +56,7 @@ public class Message implements Serializable {
     /**
      * 用于同步区块
      * */
-    public Message(String topic,String senderAddress,int height, List<Block> blockList){
+    public Message(String topic,String senderAddress,Long height, List<Block> blockList){
         this.topic = topic;
         this.senderAddress = senderAddress;
         this.height = height;
